see description
Better method for avoiding undershooting and overshooting
- [x] Gather information
- [x] Analysis and synthesis
- [x] Apply

Gather Information:
- [x]  1. Make the robot run straight, see if it actually does
It is straight enough to be considered straight
- [x]  2. Put robot in random position behind ball. Run towards the ball. Check if misses (min 10 times)
At first it was not working, soon I found out it was because the robot Was facing toward outside. The blue light was affecting the robot’s the ability to recognise the ball. Therefore I changed the robot’s direction to face towards the yellow instead. Results were successful.
- [x]  3. Make the robot run towards left, stop running when in range of shooting. Shoot
- [x]  4. ‘’	‘’	‘’	‘’	‘’	‘’	right, ‘’	‘’	‘’	‘’	‘’	‘’	‘’	‘’	‘’	‘’
- [x]  5. adjust speed, repeat step 1 and 2

Analysis and synthesis:
- [x]  1. If the robot cannot go straight, that means there exists offset in wheels
Robot goes straight
- [x]  2. See if the speed of the the robot has an effect on the “Range” of where it start shooting
For the current speed, we brake the motor for a period of time, then start shooting

Apply:
- [x]  1. Fix offset by adding offset in motor controller
No offset needed
- [x]  2. If the speed does affect the “range”, model a function and apply it into orbit.cpp
Instead of modelling a function, I made the motor brake to stop the momentum, then start shooting again.

Problem:
1. Still misses!!!! Misses approximately 1/10 times (tested with random positions)
2. Sometimes it knocks the ball out of the capture zone and has problem chasing up to it
3. Doesn’t goal track enough, most of the time it just shoot towards right of the goal and misses.